Calibration values for water quality monitoring data correction

A data set containing correction values used in practice drift corrections

data(sondeCal)

Format

a dataframe with 7 rows and 3 variables

Parameter

The parameter that you will be correcting for

'Cal Standard'

The value that the instrument should be reading for the given parameter

'Cal Value'

The value that the instrument is actually reading for the given parameter

Source

Saint Louis University Geochemistry Lab

Examples

str(sondeCal)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	7 obs. of  3 variables:
#>  $ Parameter   : chr  "SpCond" "Turbidity" "DO" "pH High" ...
#>  $ Cal Standard: int  1 0 99 10 7 1000 10
#>  $ Cal Value   : num  1.07 1.3 97.6 11.8 7.01 ...
#>  - attr(*, "spec")=List of 2
#>   ..$ cols   :List of 3
#>   .. ..$ Parameter   : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_character" "collector"
#>   .. ..$ Cal Standard: list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ Cal Value   : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_double" "collector"
#>   ..$ default: list()
#>   .. ..- attr(*, "class")= chr  "collector_guess" "collector"
#>   ..- attr(*, "class")= chr "col_spec"

head(sondeCal)

#> # A tibble: 6 x 3
#>   Parameter     `Cal Standard` `Cal Value`
#>   <chr>                  <int>       <dbl>
#> 1 SpCond                     1        1.07
#> 2 Turbidity                  0        1.3 
#> 3 DO                        99       97.6 
#> 4 pH High                   10       11.8 
#> 5 pH Low                     7        7.01
#> 6 Chloride High           1000     1411