Corrected water quality monitoring data

A data set containing corrected measurements from a YSI Sonde 6600

data(sondeClean)

Format

a dataframe with 1528 rows and 16 variables

Date

Date of measurement

Time

Time of measurement

Temp

Temperature in degrees C

SpCond

Specific conductivity raw in mS/cm

SpCond_Corr

Specific conductivity corrected in mS/cm

pH

pH raw

pH_Corr

pH corrected

pHmV

Potential reading from pH sensor in mV

Chloride

Chloride raw in mg/L

Chloride_Corr

Chloride corrected in mg/L

AmmoniumN

Ammonium-Nitrogen in mg/L

NitrateN

Nitrate-Nitrogen in mg/L

`Turbidity.`

Turbidity raw in NTU

Turbidity_Corr

Turbidity corrected in NTU

DO

Dissolved Oxygen raw in % sat

DO_Corr

Dissolved Oxygen corrected in % sat

corrFactors

A list of correction factors based on time spent in the water

Source

Saint Louis University Geochemistry Lab

Examples

str(sondeClean)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1527 obs. of  17 variables:
#>  $ Date          : chr  "9/18/2015" "9/18/2015" "9/18/2015" "9/18/2015" ...
#>  $ Time          : chr  "12:10:49" "12:15:50" "12:20:51" "12:25:51" ...
#>  $ Temp          : num  14.8 14.6 14.6 14.5 14.5 ...
#>  $ SpCond        : num  0.754 0.75 0.75 0.749 0.749 0.749 0.749 0.749 0.749 0.749 ...
#>  $ SpCond_Corr   : num  0.754 0.75 0.75 0.749 0.749 ...
#>  $ pH            : num  7.18 7.14 7.14 7.13 7.13 7.13 7.13 7.12 7.12 7.12 ...
#>  $ pH_Corr       : num  7.18 7.14 7.14 7.13 7.13 ...
#>  $ pHmV          : num  -36.4 -34.1 -33.9 -33.9 -33.6 -33.5 -33.6 -33.3 -33.3 -33.1 ...
#>  $ Chloride      : num  51.2 49.6 49.8 50.3 50.7 ...
#>  $ Chloride_Corr : num  51.2 49.6 49.7 50.3 50.7 ...
#>  $ Turbidity.    : num  3.7 -0.2 -0.1 -0.2 0 0 -0.2 -0.2 -0.2 -0.2 ...
#>  $ Turbidity_Corr: num  3.7 -0.19915 -0.0983 -0.19746 0.00339 ...
#>  $ DO            : num  92.7 93.7 94 93.2 92.7 ...
#>  $ DO_Corr       : num  92.7 93.7 93.9 93.2 92.7 ...
#>  $ corrFactors   : num  0 0.000653 0.001306 0.001956 0.002607 ...
#>  $ AmmoniumN     : num  3.35 6.29 7.84 7.67 7.13 6.49 5.78 5.32 4.89 4.6 ...
#>  $ NitrateN      : int  0 0 0 0 0 0 0 0 0 0 ...

head(sondeClean)

#> # A tibble: 6 x 17
#>   Date      Time      Temp SpCond SpCond_Corr    pH pH_Corr  pHmV Chloride
#>   <chr>     <chr>    <dbl>  <dbl>       <dbl> <dbl>   <dbl> <dbl>    <dbl>
#> 1 9/18/2015 12:10:49  14.8  0.754       0.754  7.18    7.18 -36.4     51.2
#> 2 9/18/2015 12:15:50  14.6  0.75        0.750  7.14    7.14 -34.1     49.6
#> 3 9/18/2015 12:20:51  14.6  0.75        0.750  7.14    7.14 -33.9     49.8
#> 4 9/18/2015 12:25:51  14.5  0.749       0.749  7.13    7.13 -33.9     50.3
#> 5 9/18/2015 12:30:51  14.5  0.749       0.749  7.13    7.13 -33.6     50.7
#> 6 9/18/2015 12:35:51  14.6  0.749       0.749  7.13    7.13 -33.5     50.8
#> # ... with 8 more variables: Chloride_Corr <dbl>, Turbidity. <dbl>,
#> #   Turbidity_Corr <dbl>, DO <dbl>, DO_Corr <dbl>, corrFactors <dbl>,
#> #   AmmoniumN <dbl>, NitrateN <int>