Dropping observations for equilibration period

A wrapper around dplyr::slice() for removing observations from both the head and the tail.

dr_drop(.data, head = NULL, tail = NULL)

Arguments

.data
A tbl

head
An integer >= 1 specifying the number of rows to be removed from the top of .data (or NULL)

tail
An integer >= 1 specifying the number of rows to be removed from the bottom of .data (or NULL)

Value

An object of the same class as .data with specified observations removed.

Details

When taking the instrument out of the water, there are often several observations that pass before the run can be downloaded. Additionally, once the instrument is in the water, it often takes about 30 minutes for the sensors to equilibrate. This function allows you to drop observations from the bottom and top of the dataset for each of those issues respectively.

Examples

testData <- data.frame(
   Date = c("9/18/2015", "9/18/2015", "9/18/2015", "9/18/2015", "9/18/2015", "9/18/2015"),
   Time = c("12:10:49", "12:15:50", "12:20:51", "12:25:51", "12:30:51", "12:35:51"),
   Temp = c(14.76, 14.64, 14.57, 14.51, 14.50, 14.63),
   SpCond = c(0.754, 0.750, 0.750, 0.749, 0.749, 0.749),
   stringsAsFactors = FALSE
 )

 dr_drop(testData, head = 2)

#> # A tibble: 4 x 4
#>        Date     Time  Temp SpCond
#>       <chr>    <chr> <dbl>  <dbl>
#> 1 9/18/2015 12:20:51 14.57  0.750
#> 2 9/18/2015 12:25:51 14.51  0.749
#> 3 9/18/2015 12:30:51 14.50  0.749
#> 4 9/18/2015 12:35:51 14.63  0.749

 dr_drop(testData, tail = 1)

#> # A tibble: 5 x 4
#>        Date     Time  Temp SpCond
#>       <chr>    <chr> <dbl>  <dbl>
#> 1 9/18/2015 12:10:49 14.76  0.754
#> 2 9/18/2015 12:15:50 14.64  0.750
#> 3 9/18/2015 12:20:51 14.57  0.750
#> 4 9/18/2015 12:25:51 14.51  0.749
#> 5 9/18/2015 12:30:51 14.50  0.749

 dr_drop(testData, head = 2, tail = 1)

#> # A tibble: 3 x 4
#>        Date     Time  Temp SpCond
#>       <chr>    <chr> <dbl>  <dbl>
#> 1 9/18/2015 12:20:51 14.57  0.750
#> 2 9/18/2015 12:25:51 14.51  0.749
#> 3 9/18/2015 12:30:51 14.50  0.749